WTF_web3browser/backend/requirements.txt
Flask
Flask-Cors
SQLAlchemy
psycopg2-binary
python-dotenv
flask-sqlalchemy
requests
beautifulsoup4
